Move WalletTxBuilder into seperate file
