+++ src/qt/metadexdialog.cpp
@@ -16,13 +16,15 @@
 #include "omnicore/omnicore.h"
 #include "omnicore/parse_string.h"
 #include "omnicore/pending.h"
-#include "omnicore/rules.h"
 #include "omnicore/rpctxobject.h"
+#include "omnicore/rules.h"
 #include "omnicore/sp.h"
 #include "omnicore/tally.h"
-#include "omnicore/utilsbitcoin.h"
-#include "omnicore/wallettxs.h"
 #include "omnicore/uint256_extensions.h"
+#include "omnicore/utilsbitcoin.h"
+#include "omnicore/wallettxbuilder.h"
+#include "omnicore/wallettxs.h"
+
 #include "amount.h"
 #include "sync.h"
 #include "uint256.h"
